+++ demo/simulador.py
@@ -1,48 +1,153 @@
 #!/usr/bin/env python
+# manual
 
-import cv2
-import rospy #importar ros para python
-from duckietown_msgs.msg import Twist2DStamped # importar mensajes de ROS tipo String y tipo Int32
+"""
+This script allows you to manually control the simulator or Duckiebot
+using the keyboard arrows.
+"""
+
+import time
+import websockets
+import asyncio
+import sys
+import argparse
+import pyglet
+from pyglet.window import key
+import numpy as np
+import gym
+import gym_duckietown
+from gym_duckietown.envs import DuckietownEnv
+from gym_duckietown.wrappers import UndistortWrapper
+
+# from experiments.utils import save_img
+
+parser = argparse.ArgumentParser()
+parser.add_argument('--env-name', default=None)
+parser.add_argument('--map-name', default='udem1')
+parser.add_argument('--distortion', default=False, action='store_true')
+parser.add_argument('--draw-curve', action='store_true', help='draw the lane following curve')
+parser.add_argument('--draw-bbox', action='store_true', help='draw collision detection bounding boxes')
+parser.add_argument('--domain-rand', action='store_true', help='enable domain randomization')
+parser.add_argument('--frame-skip', default=1, type=int, help='number of frames to skip')
+parser.add_argument('--seed', default=1, type=int, help='seed')
+args = parser.parse_args()
+
+if args.env_name and args.env_name.find('Duckietown') != -1:
+    env = DuckietownEnv(
+        seed = args.seed,
+        map_name = args.map_name,
+        draw_curve = args.draw_curve,
+        draw_bbox = args.draw_bbox,
+        domain_rand = args.domain_rand,
+        frame_skip = args.frame_skip,
+        distortion = args.distortion,
+    )
+else:
+    env = gym.make(args.env_name)
+
+env.reset()
+env.render()
+
+@env.unwrapped.window.event
+def on_key_press(symbol, modifiers):
+    """
+    This handler processes keyboard commands that
+    control the simulation
+    """
+
+    if symbol == key.BACKSPACE or symbol == key.SLASH:
+        print('RESET')
+        env.reset()
+        env.render()
+    elif symbol == key.PAGEUP:
+        env.unwrapped.cam_angle[0] = 0
+    elif symbol == key.ESCAPE:
+        env.close()
+        sys.exit(0)
+
+    # Take a screenshot
+    # UNCOMMENT IF NEEDED - Skimage dependency
+    # elif symbol == key.RETURN:
+    #     print('saving screenshot')
+    #     img = env.render('rgb_array')
+    #     save_img('screenshot.png', img)
+
+# Register a keyboard handler
+key_handler = key.KeyStateHandler() 
+env.unwrapped.window.push_handlers(key_handler)
+
+actionMsg = 'Connected!'
+def update():
+    """
+    This function is called at every frame to handle
+    movement/stepping and redrawing
+    """
+
+    global actionMsg
+    action = np.array([0.0, 0.0])
+
+    if key_handler[key.UP]:
+        action = np.array([0.44, 0.0])
+    if key_handler[key.DOWN]:
+        action = np.array([-0.44, 0])
+    if key_handler[key.LEFT]:
+        action = np.array([0.35, +1])
+    if key_handler[key.RIGHT]:
+        action = np.array([0.35, -1])
+    if key_handler[key.SPACE]:
+        action = np.array([0, 0])
+
+    # Speed boost
+    if key_handler[key.LSHIFT]:
+        action *= 1.5
+
+    obs, reward, done, info = env.step(action)
+    print('step_count = %s, reward=%.3f' % (env.unwrapped.step_count, reward))
+
+    if key_handler[key.RETURN]:
+        from PIL import Image
+        im = Image.fromarray(obs)
+
+        im.save('screen.png')
+
+    lineal = str(action[0])
+    angular = str(action[1])
+    actionMsg = (lineal+","+angular)
+    if done:
+        print('done!')
+        env.reset()
+        env.render()
+
+    env.render()
+
+async def connect(websocket, path):
+    delta_time = 1.0 / env.unwrapped.frame_rate
+    while keepRunning:
+        start = time.time()
+        await websocket.send(actionMsg)
+        spent = time.time() - start
+        wait = delta_time - spent
+        await asyncio.sleep(wait)
 
 
-class Simulador(object):
-	def __init__(self, args):
-		super(Simulador, self).__init__()
-		self.args = args
-		self.publisher = rospy.Publisher("/duckiebot/wheels_driver_node/car_cmd", Twist2DStamped, queue_size=10)
-		self.twist = Twist2DStamped()
-		
-		nothing = lambda x: x
+async def startSimulator():
+    delta_time = 1.0 / env.unwrapped.frame_rate
+    
 
-		cv2.namedWindow('Simulador')
-		cv2.createTrackbar('Linear','Simulador',100,200,nothing)
-		cv2.createTrackbar('Angular','Simulador',100,200,nothing)
+    while keepRunning:
+        start = time.time()
+        update()
+        spent = time.time() - start
 
+        wait = delta_time - spent
+        await asyncio.sleep(wait) 
 
-	def actualizar(self):
-		while 1:
-			k = cv2.waitKey(1)
-			if k == 27:
-				break
+loop  = asyncio.get_event_loop()
+keepRunning = True
+start_server = websockets.serve(connect, "localhost", 8765)
+asyncio.ensure_future(start_server)
+asyncio.ensure_future(startSimulator())
 
-			linear = cv2.getTrackbarPos('Linear','Simulador') - 100
-			angular = cv2.getTrackbarPos('Angular','Simulador') - 100
-
-			linear = linear / 100.0
-			angular = angular / -12.5
-
-
-			self.twist.v = linear
-			self.twist.omega = angular
-			
-			self.publisher.publish(self.twist)
-
-def main():
-	rospy.init_node('simulador') #creacion y registro del nodo!
-
-	obj = Simulador('args') # Crea un objeto del tipo Simulador, cuya definicion se encuentra arriba
-
-	obj.actualizar() #llama al metodo publicar del objeto obj de tipo Simulador
-
-if __name__ =='__main__':
-	main()
+loop.run_forever()
+keepRunning = False
+env.close()
